requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.5
pandas==1.2.1
Pillow==8.1.0
praw==7.1.0
prawcore==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
squarify==0.4.3
tqdm==4.56.0
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0